Fix issue with company addresses / refactoring
