+++ skins/laika/src/store/address/constants.ts
@@ -1,12 +1,17 @@
 import { AddressTypeModel } from '@/view_models/AddressTypeModel';
 
-export const REQUIRED_FIELDS: {[key: number]: string[]} = {
+/** AddressTypeModel: String array of required fields **/
+export interface AddressRequirements {
+	[ key: number ]: string[]
+}
+
+export const REQUIRED_FIELDS: AddressRequirements = {
 	[ AddressTypeModel.PERSON ]: [ 'salutation', 'firstName', 'lastName', 'street', 'postcode', 'city', 'email' ],
 	[ AddressTypeModel.COMPANY ]: [ 'companyName', 'street', 'postcode', 'city', 'email' ],
 	[ AddressTypeModel.ANON ]: [],
 };
 
-export const REQUIRED_FIELDS_ADDRESS_UPDATE: {[key: number]: string[]} = {
+export const REQUIRED_FIELDS_ADDRESS_UPDATE: AddressRequirements = {
 	[ AddressTypeModel.PERSON ]: [ 'salutation', 'firstName', 'lastName', 'street', 'postcode', 'city' ],
 	[ AddressTypeModel.COMPANY ]: [ 'companyName', 'street', 'postcode', 'city' ],
 };
